+++ frameworks/qwik/src/components/Form/Form.tsx
@@ -23,7 +23,9 @@ export type FormProps<TSchema extends Schema = Schema> = Omit<
   /**
    * The submit handler called when the form is submitted and validation succeeds.
    */
-  readonly onSubmit$: QRL<SubmitHandler<TSchema> | SubmitEventHandler<TSchema>>;
+  readonly onSubmit$:
+    | QRL<SubmitHandler<TSchema>>
+    | QRL<SubmitEventHandler<TSchema>>;
 };
 
 /**