+++ playgrounds/qwik/src/routes/payment/index.tsx
@@ -58,7 +58,7 @@ export default component$(() => {
     <Form
       of={paymentForm}
       class="space-y-12 md:space-y-14 lg:space-y-16"
-      onSubmit$={(output) => console.log(output)}
+      onSubmit$={(output, _) => console.log(output)}
     >
       <FormHeader of={paymentForm} heading="Payment form" />
       <div class="space-y-8 md:space-y-10 lg:space-y-12">