+++ playgrounds/qwik/src/routes/todos/index.tsx
@@ -65,7 +65,7 @@ export default component$(() => {
     <Form
       of={todoForm}
       class="space-y-12 md:space-y-14 lg:space-y-16"
-      onSubmit$={(output) => console.log(output)}
+      onSubmit$={(output, _) => console.log(output)}
     >
       <FormHeader of={todoForm} heading="Todo form" />