+++ playgrounds/react/src/routes/nested/index.tsx
@@ -50,7 +50,7 @@ export default function NestedPage() {
     <Form
       of={nestedForm}
       className="space-y-12 md:space-y-14 lg:space-y-16"
-      onSubmit={(output) => console.log(output)}
+      onSubmit={(output, _) => console.log(output)}
     >
       <FormHeader of={nestedForm} heading="Nested form" />