+++ playgrounds/react/src/routes/todos/index.tsx
@@ -58,7 +58,7 @@ export default function Page() {
     <Form
       of={todoForm}
       className="space-y-12 md:space-y-14 lg:space-y-16"
-      onSubmit={(output) => console.log(output)}
+      onSubmit={(output, _) => console.log(output)}
     >
       <FormHeader of={todoForm} heading="Todo form" />