+++ playgrounds/solid/src/routes/todos/index.tsx
@@ -60,7 +60,7 @@ export default function TodosPage() {
       <Form
         of={todoForm}
         class="space-y-12 md:space-y-14 lg:space-y-16"
-        onSubmit={(output) => console.log(output)}
+        onSubmit={(output, _) => console.log(output)}
       >
         <FormHeader of={todoForm} heading="Todo form" />