+++ playgrounds/svelte/src/routes/nested/+page.svelte
@@ -49,7 +49,7 @@
   <title>Nested form</title>
 </svelte:head>
 
-<Form of={nestedForm} onsubmit={(output) => console.log(output)}>
+<Form of={nestedForm} onsubmit={(output, _) => console.log(output)}>
   <div class="space-y-12 md:space-y-14 lg:space-y-16">
     <FormHeader of={nestedForm} heading="Nested form" />