+++ playgrounds/vue/src/views/NestedView.vue
@@ -43,7 +43,7 @@ const nestedForm = useForm({
   <Form
     :of="nestedForm"
     class="space-y-12 md:space-y-14 lg:space-y-16"
-    @submit="(output) => console.log(output)"
+    @submit="(output, _) => console.log(output)"
   >
     <FormHeader :of="nestedForm" heading="Nested form" />