+++ playgrounds/vue/src/views/PaymentView.vue
@@ -58,7 +58,7 @@ const type = computed(() => getInput(paymentForm, { path: ['type'] }));
   <Form
     :of="paymentForm"
     class="space-y-12 md:space-y-14 lg:space-y-16"
-    @submit="(output) => console.log(output)"
+    @submit="(output, _) => console.log(output)"
   >
     <FormHeader :of="paymentForm" heading="Payment form" />
     <div class="space-y-8 md:space-y-10 lg:space-y-12">