+++ playgrounds/vue/src/views/TodosView.vue
@@ -56,7 +56,7 @@ const todoForm = useForm({
   <Form
     :of="todoForm"
     class="space-y-12 md:space-y-14 lg:space-y-16"
-    @submit="(output) => console.log(output)"
+    @submit="(output, _) => console.log(output)"
   >
     <FormHeader :of="todoForm" heading="Todo form" />