+++ website/src/routes/playground/login/index.tsx
@@ -42,7 +42,7 @@ export default component$(() => {
     <Form
       of={loginForm}
       class="flex flex-col gap-12 md:gap-14 lg:gap-16"
-      onSubmit$={(output) => alert(JSON.stringify(output, null, 2))}
+      onSubmit$={(output, _) => alert(JSON.stringify(output, null, 2))}
     >
       <FormHeader of={loginForm} heading="Login form" />
       <div class="flex flex-col gap-8 md:gap-10 lg:gap-12">